Fixed more TODOs

+++ script/vue/filesystem.js
@@ -23,17 +23,26 @@ Vue.component('filesystem', {
       this.opened = null;
     },
     addStore: function(fileHandle) {
-      this.filesystem.addStore(fileHandle);
+      if (this.filesystem instanceof LocalFilesystem) {
+        this.filesystem.addStore(fileHandle);
+      }
     },
     removeStore: function(store) {
-      this.filesystem.removeStore(store.id);
+      if (this.filesystem instanceof LocalFilesystem) {
+        this.filesystem.removeStore(store.id);
+      }
     },
     sync: function() {
-      this.filesystem.refreshStores();
+      if (this.filesystem instanceof RemoteFilesystem) {
+        this.filesystem.refreshStores();
+      }
     }
   },
+  mounted: function() {
+    setTimeout(() => this.sync(), 500);
+  },
   template: `
-    <div class="card">
+    <div class="card mb-2">
       <div class="card-body">
         <h5 class="card-title">{{ remote ? 'Remote Filesystem' : 'Local Filesystem' }}</h5>
         <stores-list v-bind:stores="filesystem.stores"
